requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
pyyaml==6.0.2
SQLAlchemy==2.0.36
